Earlier simplification of byte compilation code led to junk directories
An earlier check-in removed a special case for byte-compiling: source
files from the filesystem and source files from zipfiles are treated
the same. Sadly enough this used the code for zipped sources, which as
it turns out is broken and leads to junk directories on the fileystem.

This check-in fixes that problem.