Erase build and dist directories before builds
With this patch py2app will remove previous builds before attempting
a new one. This is needed because the results of a previous
build can cause problems with a new build, and keeping the
previous build around is not usefull anyway because py2app will
do a full rebuild anyway.