Issue #15: Add an option for initializing os.environ as if the application is started from the Terminal
With option --emulate-shell-environment py2app will generate an application
bundle that updates os.environ using values gotten from a fresh login-shell.
This way os.environ will mostly contain the same data as can be seen in
a shell window in the Terminal.

This is not ideal though, and I wouldn't use this option unless there is a
very clear requirement for doing so. In particular this won't see environment
updates done in an interactive shell, which might still confuse users.