Fix for issue #16: Ensure that argv_emulation works with 64-bit code and python 3.x
This check-in replaces the argv emulation code with a version that uses
ctypes to call non-deprecated Carbon/CoreFoundation APIs. This makes it
possible to use the argv emulator in 64-bit code and with Python 3.x.

NOTE: The emulator is effectively broken with python 3.x at the moment
unless you're using an intel-only build because the wrapper code needs
to be rebuild.