(temporarily) add recipe for pyzmq
Pyzmq has a package 'zmq' that contains a dylib that is used
by extensions. Py2app currently copies the dylib into
site-packages.zip, where it cannot be found by dyld and hence
the app bundle is non-functional.

The recipe ensures that zmq is included outsite of site-packages.zip,
hence avoiding the bug in py2app.