Issue 59: app bundle is changed on first launch
With this patch site.py is included as 'site.pyc' (or 'site.pyo')
to avoid that site.py is compiled on first launch.

The testsuite now verifies that the app bundle does not changed
when the application is launched.

The issue also mentions changing other files not included in the zipfile
that get compiled at launch, that probably caused by an issue I
fixed earlier. Test_package_data verifies that files aren't changed
and passed without further changes to py2app.