Add an incomplete pyenchant recipe.
Related to issue #37

The recipe is incomplete because it works only
when the enchant C library is embedded in the package,
not when using an enchant C library outside of the
package.

Support the latter likely requires adding a prescript
(setting environment variables) and a way to explictly
include shared libraries (e.g. force py2app to include
libenchant.dylib, its dependencies and data files)