Always include pkg_resources because of an implicit import in namespace package
<http://packages.python.org/distribute/setuptools.html#namespace-packages>
describes how to create a namespace package with setuptools and distribute.
The documentation says that the __init__.py must contain:

   __import__('pkg_resources').declare_namespace(__name__)

This import of pkg_resources isn't recognized by the bytecode scanner,
therefore include pkg_resources unconditionally.

NOTE: In the longer run it would be worthwhile to use an AST visitor to detect
calls to __import__ with constant arguments.