Add --force-system-tk option to py2app (issue #92)
This changeset adds an option '--force-system-tk' that will ensure
that _tkinter.so is linked to the system version of Tcl/Tk
(/System/Library/Frameworks/...) instead of some other framework.

Causes a build error when Tkinter is linked against a non-compatible
version (not a framework, or a framework version not present in
/System/Library/Frameworks).

Also remove duplicate definitions of a number of methods (!) and
remove a block of code that was disabled with 'if 0'.

Closes issue #92